My items added & design customised

+++ src/Components/ManageItem.jsx
@@ -1,23 +1,36 @@
 import React from 'react';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faTrashCan, faLongArrowAltRight, faXmark } from '@fortawesome/free-solid-svg-icons'
+import { useNavigate } from 'react-router-dom';
+import { faTrashCan } from '@fortawesome/free-solid-svg-icons'
 import axios from 'axios';
+import { toast } from 'react-toastify';
 
 
 
-const ManageItem = ({product,products,setProducts}) => {
-    console.log(setProducts);
+const ManageItem = ({ product, products, setProducts }) => {
     const { _id, name, picture, description, price, quantity, supplier } = product;
-    
+    const navigate = useNavigate()
+
     const handleDeleteItem = (id) => {
-        console.log(products.filter(fproduct => fproduct._id !== id));
-        (async () => {
-            const { data } = await axios.delete(`https://powerful-woodland-06362.herokuapp.com/product/${id}`);
-            if(data){
-               const remaningProducts =  products.filter(fproduct => fproduct._id !== id);
-               setProducts(remaningProducts)
-            }
-          })()
+        const realy = window.confirm('Are you sure?');
+        if (realy) {
+            (async () => {
+                const { data } = await axios.delete(`https://powerful-woodland-06362.herokuapp.com/product/${id}`);
+                if (data) {
+                    toast.success('Item Deleted', {
+                        position: "top-right",
+                        autoClose: 1000,
+                        hideProgressBar: true,
+                        closeOnClick: true,
+                        pauseOnHover: true,
+                        draggable: true,
+                        progress: undefined,
+                        });
+                    const remaningProducts = products.filter(fproduct => fproduct._id !== id);
+                    setProducts(remaningProducts);
+                }
+            })();
+        }
     }
     return (
         <>
@@ -34,11 +47,14 @@ const ManageItem = ({product,products,setProducts}) => {
                 <td className="md:px-6 py-4">
                     {quantity}
                 </td>
-                <td className="md:px-6 py-4">
+                <td className="md:pl-6 py-4">
                     {price}
                 </td>
-                <td className="md:px-6 py-4 text-right">
-                <FontAwesomeIcon onClick={()=>handleDeleteItem(_id)} className='w-6 h-6  cursor-pointer  text-red-600' icon={faTrashCan}></FontAwesomeIcon>
+                <td className="md:pl-6 py-4">
+                <button onClick={()=>navigate(`/inventory/${_id}`)}  className=' rounded border border-rose-500 px-2 py-1 hover:bg-rose-500 hover:text-white focus:right-1'>stock update</button>
+                </td>
+                <td className="md:px-2 py-4 text-right ">
+                    <FontAwesomeIcon onClick={() => handleDeleteItem(_id)} className='w-6 h-6  cursor-pointer  text-red-600 active:bg-rose-300 p-2 rounded-full' icon={faTrashCan}></FontAwesomeIcon>
                 </td>
             </tr>
         </>
